+++ src/torManager/torManager.test.ts
@@ -1,13 +1,25 @@
 /* eslint import/first: 0 */
 import { Tor } from './torManager'
-import { ZBAY_DIR_PATH } from '../constants'
 import { getPorts } from '../utils'
-import { spawnTorProcess } from '..//testUtils'
+import { createTmpDir, spawnTorProcess, TmpDir, tmpZbayDirPath } from '../testUtils'
 
 jest.setTimeout(30_000)
 
+let tmpDir: TmpDir
+let tmpAppDataPath: string
+
+beforeEach(() => {
+  jest.clearAllMocks()
+  tmpDir = createTmpDir()
+  tmpAppDataPath = tmpZbayDirPath(tmpDir.name)
+})
+
+afterEach(async () => {
+  tmpDir.removeCallback()
+})
+
 test('start and close tor', async () => {
-  const tor = await spawnTorProcess()
+  const tor = await spawnTorProcess(tmpAppDataPath)
   await tor.init()
   await tor.kill()
 })
@@ -17,14 +29,14 @@ test('start tor, do not kill tor process and start again', async () => {
   const ports = await getPorts()
   const libPath = `${process.cwd()}/tor`
   const tor = new Tor({
-    appDataPath: ZBAY_DIR_PATH,
+    appDataPath: tmpAppDataPath,
     socksPort: ports.socksPort,
     torPath: torPath,
     controlPort: ports.controlPort,
     options: {
       env: {
         LD_LIBRARY_PATH: libPath,
-        HOME: ZBAY_DIR_PATH
+        HOME: tmpAppDataPath
       },
       detached: true
     }
@@ -33,14 +45,14 @@ test('start tor, do not kill tor process and start again', async () => {
   await tor.init()
 
   const torSecondInstance = new Tor({
-    appDataPath: ZBAY_DIR_PATH,
+    appDataPath: tmpAppDataPath,
     socksPort: ports.socksPort,
     torPath: torPath,
     controlPort: ports.controlPort,
     options: {
       env: {
         LD_LIBRARY_PATH: libPath,
-        HOME: ZBAY_DIR_PATH
+        HOME: tmpAppDataPath
       },
       detached: true
     }
@@ -50,7 +62,7 @@ test('start tor, do not kill tor process and start again', async () => {
 })
 
 test('spawn new hidden service', async () => {
-  const tor = await spawnTorProcess()
+  const tor = await spawnTorProcess(tmpAppDataPath)
   await tor.init()
   const hiddenService = await tor.createNewHiddenService(4343, 4343)
   expect(hiddenService.onionAddress).toHaveLength(56)
@@ -58,7 +70,7 @@ test('spawn new hidden service', async () => {
 })
 
 test('spawn hidden service using private key', async () => {
-  const tor = await spawnTorProcess()
+  const tor = await spawnTorProcess(tmpAppDataPath)
   await tor.init()
   const hiddenServiceOnionAddress = await tor.spawnHiddenService({
     virtPort: 4343,
@@ -68,4 +80,13 @@ test('spawn hidden service using private key', async () => {
   })
   expect(hiddenServiceOnionAddress).toBe('u2rg2direy34dj77375h2fbhsc2tvxj752h4tlso64mjnlevcv54oaad')
   await tor.kill()
+})
+
+test('generate hashed password', async () => {
+  const tor = await spawnTorProcess(tmpAppDataPath)
+  tor.generateHashedPassword()
+  console.log(tor.torHashedPassword)
+  console.log(tor.torPassword)
+  expect(tor.torHashedPassword).toHaveLength(62)
+  expect(tor.torPassword).toHaveLength(32)
 })